Implement two step workflow
First you run the subcommand from-linear and rerun as many times as you
need to fetch every linear issue into the corpus and then you run
to-github to export them all to github. You can ctrl+c and rerun the
from-linear at any point but ideally you want to-github to run
continously until complete.

Linear's rate limits are too low. It take's a while to fetch every issue
and managing things manually with $BYELINEAR_BEFORE and time.Sleep was a
pain.

And if at any point something goes wrong and say there's a bug in the
exporter, well now the whole process has to be restarted.